refactor: update based on PR comments
