+++ src/main/proto/wfa/measurement/queue/testing/BUILD.bazel
@@ -1,11 +1,13 @@
 load("@rules_proto//proto:defs.bzl", "proto_library")
 load(
     "@wfa_rules_kotlin_jvm//kotlin:defs.bzl",
-    "kt_jvm_grpc_proto_library",
     "kt_jvm_proto_library",
 )
 
-package(default_visibility = ["//visibility:public"])
+package(
+    default_testonly = True,
+    default_visibility = ["//visibility:public"],
+)
 
 IMPORT_PREFIX = "/src/main/proto"
