+++ src/main/proto/wfa/measurement/queue/testing/test_work.proto
@@ -14,12 +14,12 @@
 
 syntax = "proto3";
 
-package org.wfa.measurement.queue;
+package org.wfa.measurement.queue.testing;
 
 import "google/api/field_behavior.proto";
 import "google/api/resource.proto";
 
-option java_package = "org.wfa.measurement.queue";
+option java_package = "org.wfa.measurement.queue.testing";
 option java_multiple_files = true;
 option java_outer_classname = "TestWorkProto";
