+++ src/test/kotlin/org/wfanet/measurement/gcloud/pubsub/BUILD.bazel
@@ -13,7 +13,7 @@ kt_jvm_test(
         "//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub:subscriber",
         "//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub/testing:google_pub_sub_emulator_client",
         "//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub/testing:google_pub_sub_emulator_provider",
-        "//src/main/proto/wfa/measurement/queue:test_work_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/queue/testing:test_work_kt_jvm_proto",
     ],
 )
 
@@ -31,6 +31,6 @@ kt_jvm_test(
         "//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub:subscriber",
         "//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub/testing:google_pub_sub_emulator_client",
         "//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub/testing:google_pub_sub_emulator_provider",
-        "//src/main/proto/wfa/measurement/queue:test_work_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/queue/testing:test_work_kt_jvm_proto",
     ],
 )
